Rename PotentialFunction => Potential
