+++ examples/custom-potential.rs
@@ -15,8 +15,8 @@ struct LJ {
     b: f64
 }
 
-/// All we need to do is to implement the PotentialFunction trait
-impl PotentialFunction for LJ {
+/// All we need to do is to implement the Potential trait
+impl Potential for LJ {
     /// The energy function give the energy at distance `r`
     fn energy(&self, r: f64) -> f64 {
         self.a / r.powi(12) - self.b / r.powi(6)
